Refactor common test tooling into testtools. Skip tests that currently fail on golang version. Now the game is to make them pass.
