+++ dvol_python/test_dvol.py
@@ -10,18 +10,17 @@ from hypothesis.strategies import binary, characters, dictionaries, sets, text
 from twisted.trial.unittest import TestCase
 from twisted.python.filepath import FilePath
 from twisted.python.usage import UsageError
-from unittest import skipIf
 import subprocess
 import os
 import json
 
-TEST_DVOL_BINARY = os.environ.get("TEST_DVOL_BINARY", False)
+from testtools import (
+    CalledProcessErrorWithOutput, TEST_DVOL_BINARY, skip_if_go_version,
+    skip_if_python_version
+)
+
 DVOL_BINARY = os.environ.get("DVOL_BINARY", "./dvol")
 ARGS = ["--disable-docker-integration"]
-
-
-class CalledProcessErrorWithOutput(Exception):
-    pass
 
 if TEST_DVOL_BINARY:
     # Test an alternative implementation of dvol, such as one available as a
@@ -92,17 +91,6 @@ def path_segments():
 
 volume_names = path_segments
 branch_names = path_segments
-
-skip_if_go_version = skipIf(
-    TEST_DVOL_BINARY,
-    "Not expected to work in go version"
-)
-
-skip_if_python_version = skipIf(
-    not TEST_DVOL_BINARY,
-    "Not expected to work in Python version"
-)
-
 
 class VoluminousTests(TestCase):
     def setUp(self):
